Library/fips: minor fix

+++ runtest.sh
@@ -38,22 +38,25 @@ rlJournalStart
     if ! [ -e /var/tmp/fips-reboot ]; then
         rlPhaseStartTest
 
-            rlRun "fipsIsSupported" || rlDie
+            # Check that FIPS 140 mode is supported.
+            rlRun "fipsIsSupported" 0,1
+            if [ $? -eq 0 ]; then
 
-            # Initially, FIPS mode is disabled.
-            rlRun "fipsIsEnabled" 1
+                # Initially, FIPS mode is disabled.
+                rlRun "fipsIsEnabled" 1
+                
+                # Enable it.
+                rlRun "fipsEnable" 0
+                
+                # Before completing setup by restart, system is misconfigured.
+                rlIsRHEL ">6.5" && rlRun "fipsIsEnabled" 2
+                
+                rlRun "touch /var/tmp/fips-reboot" 0
+                
+            rlPhaseEnd
 
-            # Enable it.
-            rlRun "fipsEnable" 0
-            
-            # Before completing setup by restart, system is misconfigured.
-            rlIsRHEL ">6.5" && rlRun "fipsIsEnabled" 2
-            
-            rlRun "touch /var/tmp/fips-reboot" 0
-
-        rlPhaseEnd
-
-        rhts-reboot
+            rhts-reboot
+            fi
     else
         rlPhaseStartTest
